pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drift-lens"
version = "0.1.0"
description = "Your metrics lie. Your embeddings don't. Detect embedding space drift before accuracy drops."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Praful Reddy", email = "prafulreddy@example.com"},
]
keywords = ["embeddings", "drift-detection", "monitoring", "mlops", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=12.0",
    "scikit-learn>=1.3",
    "umap-learn>=0.5",
    "ripser>=0.6",
    "plotly>=5.0",
    "streamlit>=1.28",
    "click>=8.0",
    "scipy>=1.11",
]

[project.optional-dependencies]
torch = ["torch>=2.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.urls]
Homepage = "https://github.com/PRAFULREDDYM/drift-lens"
Repository = "https://github.com/PRAFULREDDYM/drift-lens"
Issues = "https://github.com/PRAFULREDDYM/drift-lens/issues"

[project.scripts]
drift-lens = "drift_lens.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
